Update /test/integration/rest_api/project.js to use PageObject pattern
- Move expect statements to matchers, add additional page objects, rename test-integration to test-e2e
