+++ assets/test/e2e/matchers/expectDescription.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports =  function(description, className) {
+  expect(element(by.css(className || '.project-description')).getText()).toEqual(description);
+};
