+++ assets/test/e2e/matchers/expectDisplayName.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(displayName, className) {
+  expect(element(by.css(className || '.project-display-name')).getText()).toEqual(displayName);
+};
