+++ assets/test/e2e/matchers/expectHeading.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(text, heading) {
+  expect(element(by.cssContainingText(heading || 'h1', text)).isPresent()).toBe(true);
+};
