+++ assets/test/e2e/matchers/expectInput.js
@@ -1,0 +1,6 @@
+'use strict';
+
+// by.model is the ng-model attribute
+module.exports = function(model) {
+  expect(element(by.model(model)).isPresent()).toBe(true);
+};
