+++ assets/test/e2e/matchers/expectName.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(name, className) {
+  expect(element(by.css(className || '.project-name')).getText()).toEqual(name);
+};
