+++ assets/test/e2e/page_objects/servicesPage.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var browse = require('../helpers/browser.js');
+
+exports.visit = function(project) {
+  return browse.goTo('/project/' + project.name + '/browse/services');
+};
