+++ assets/test/protractor-chrome.conf.js
@@ -89,8 +89,10 @@ exports.config = {
   // run.
   suites: {
     // smoke: 'spec/smoketests/*.js',
-    rest_api: 'integration/rest_api/*.js', // This suite of tests should only require a running master api, it should not require a node
-    e2e: 'integration/e2e.js'
+    // TODO: naming.
+    rest_api: 'e2e/rest_api/*.js', // This suite of tests should only require a running master api, it should not require a node
+    // TODO: naming.
+    e2e: 'e2e/e2e.js'
   },
 
   // ---------------------------------------------------------------------------
@@ -222,10 +224,10 @@ exports.config = {
       baseDirectory: './test/tmp/screenshots',
       takeScreenShotsOnlyForFailedSpecs: true,
       pathBuilder: function pathBuilder(spec, descriptions, results, capabilities) {
-        // Return '<specname>' as path for screenshots: 
-        // Example: 'list-should work'. 
+        // Return '<specname>' as path for screenshots:
+        // Example: 'list-should work'.
         return descriptions.reverse().join(' ');
-      }       
+      }
     }));
   },
 
@@ -336,4 +338,3 @@ exports.config = {
     format: 'summary'
   }
 };
-
